test and fix for prefilled mass message
I fleshed out the feature test for client statuses to assert that
prefilled mass messages spawned by a link in the client status banner
worked. The functionality was broken, so I fixed it.

I also renamed a lot of things to reflect that we're using reporting
relationships instead of clients in various places.

[#156680889]